refactor(testutil): split god-utility into 5 focused files (#156)
Closes manchtools/power-manage-server#156 (audit F039).

The 744-LOC internal/testutil/testutil.go is split along organisational
seams that mirror how callers use these helpers:

- testutil.go (44 LOC) — package vars, init, NewID
- postgres.go (103 LOC) — testcontainer setup
- factories_user.go (224 LOC) — users, roles, user groups, contexts, JWT
- factories_device.go (100 LOC) — devices, device groups, assignments
- factories_action.go (190 LOC) — actions, action sets, definitions, assignments, tokens
- factories_idp.go (170 LOC) — identity providers, SCIM, TOTP, Encryptor

All files are now under 300 LOC (issue's acceptance criterion). Every
public symbol stays on package testutil so call sites are unchanged
— this is purely a file-shuffle, no API changes, no behaviour changes.
The package doc comment in testutil.go is the new index for what
lives where.

Verified: `go build ./...`, `go vet ./...` clean.
